:sparkles: Feat: sort suggestion & add docs
