+++ src/components/Suggestions/utils.ts
@@ -22,3 +22,22 @@ export function matchIndexToPosition(value: string, index: number): {lineNumber:
     }
     return {lineNumber: lineNumber, column: columnNumber};
 }
+
+export function getCursorPositionClause(value: string, cursorPosition: {lineNumber: number, column: number}): {clause: string, index: number} {
+    // Get the current clause based on the cursor position, or START if no clause is found or the cursor is at
+    // the start of the query (; or ()). The regex matches the last occurrence of a clause keyword before the cursor.
+    const clausePattern = new RegExp("(select|use|from|where|group by|order by|\\s\\(|;)(?![\\s\\S]*(select|use|from|where|group by|order by|\\s\\(|;))", "i");
+    const cursorIndex = positionToIndex(value, cursorPosition);
+    const match = clausePattern.exec(value.substring(0, cursorIndex));
+    if (match !== null) {
+        const clause = match[1].includes("(") || match[1].includes(";") ? "START" : match[1];
+        return {
+            clause: clause,
+            index: match.index
+        };
+    }
+    return {
+        clause: "START",
+        index: 0
+    };
+}
